+++ Menu.lua
@@ -5,7 +5,7 @@ local panelData = {
     name = "WarCry",
     displayName = "WarCry",
     author = "|cFFC0CBm00ny|r",
-    version = "0.2.0",
+    version = "0.3.0",
     website = "https://www.esoui.com/downloads/fileinfo.php?id=3191",
     slashCommand = "/wcsettings",	--(optional) will register a keybind to open to this panel
     registerForRefresh = true,	--boolean (optional) (will refresh all options controls when a setting is changed and when the panel is shown)
@@ -15,21 +15,18 @@ local panelData = {
 local optionsTable = {
     [1] = {
         type = "header",
-        name = "general settings",
+        name = "group control settings",
         width = "full",	--or "half" (optional)
     },
     [2] = {
-        type = "slider",
-        name = "Collectible ID",
-        tooltip = "ID of the collectible that you want to play",
-        min = 1,
-        max = 1000,
-        disabled = true,
-        step = 1,	--(optional)
-        getFunc = function() return WarCry.savedVariables.collectibleID end,
-        setFunc = function(value) WarCry.savedVariables.collectibleID = value end,
+        type = "checkbox",
+        name = "synced WarCry",
+        tooltip = "enables/disables listening for chat events to sync warcry in the group",
+        getFunc = function() return WarCry.savedVariables.groupCommands end,
+        setFunc = function(value) WarCry.toggleListener(value) end,
         width = "half",	--or "half" (optional)
-        default = 353,	--(optional)
+        default = true,
+        --warning = "Will need to reload the UI.",	--(optional)
     },
     [3] = {
         type = "texture",
@@ -39,22 +36,7 @@ local optionsTable = {
         imageHeight = 64,	--max of 100
         width = "half",	--or "half" (optional)
     },
-    [4] = {
-        type = "header",
-        name = "group control settings",
-        width = "full",	--or "half" (optional)
-    },
-    [5] = {
-        type = "checkbox",
-        name = "synced warcry",
-        tooltip = "enables/disables listening for chat events to sync warcry in the group",
-        getFunc = function() return WarCry.savedVariables.groupCommands end,
-        setFunc = function(value) WarCry.toggleListener(value) end,
-        width = "full",	--or "half" (optional)
-        default = true,
-        --warning = "Will need to reload the UI.",	--(optional)
-    },
 }
 
-LAM2:RegisterAddonPanel("WarCry", panelData)
-LAM2:RegisterOptionControls("WarCry", optionsTable)
+LAM2:RegisterAddonPanel("WarCryMenu", panelData)
+LAM2:RegisterOptionControls("WarCryMenu", optionsTable)