+++ WarCry.lua
@@ -1,24 +1,16 @@
-WarCry = {}
-WarCry.name = "WarCry"
-WarCry.color = "8B0000"
-WarCry.credits = "@m00nyONE"
-WarCry.slashCmd = "/wc"
-WarCry.variableVersion = 1
-WarCry.listenString = "ahu"
-WarCry.defaultVariables = {
-    collectibleID = 353,
-    groupCommands = true,
-}
+WarCry = WarCry or {}
+WarCry.ahuListenString = "ahu"
+WarCry.ahuCollectibleIDs = {353}
+WarCry.roarListenSTring = "roar"
+WarCry.roarCollectibleIDs = {9432, 9590}
 
-ZO_CreateStringId("SI_BINDING_NAME_WARCRY_PLAY", "Play Warcry")
-ZO_CreateStringId("SI_BINDING_NAME_WARCRY_GROUP_PLAY", "Play Group Warcry")
 
--- currently fixed
-function WarCry.Play()
-    UseCollectible(WarCry.savedVariables.collectibleID)
-end
+ZO_CreateStringId("SI_BINDING_NAME_WARCRY_AHU_PLAY", "Play AHU Warcry")
+--- ZO_CreateStringId("SI_BINDING_NAME_WARCRY_AHU_GROUP_PLAY", "Play AHU Group Warcry")
+ZO_CreateStringId("SI_BINDING_NAME_WARCRY_ROAR_PLAY", "Play ROAR Warcry")
+--- ZO_CreateStringId("SI_BINDING_NAME_WARCRY_ROAR_GROUP_PLAY", "Play ROAR Group Warcry")
 
--- initiates a synced warcry event
+--[[ -- initiates a synced warcry event
 function WarCry.GroupPlay()
 
     -- check if player is in a group
@@ -33,45 +25,7 @@ function WarCry.GroupPlay()
     end
     d("|c8B0000WarCry: you are not in a group|r")
     return
-end
-
-
-
-WarCry.chatCallback = function(_, messageType, _, message, _, fromDisplayName)
-    -- check if message is in group chat
-    if messageType == CHAT_CHANNEL_PARTY then
-        -- check if message is the one we are listening for
-        if string.lower(message) == string.lower(WarCry.listenString) then
-            -- check if the iniciator is group leader
-            if fromDisplayName == GetUnitDisplayName(GetGroupLeaderUnitTag()) then
-                WarCry.Play()
-            end
-        end
-    end
-end
-
-
-function WarCry.toggleListener(value)
-    -- toggle listener
-    if value then
-        WarCry.startChatListener()
-    else
-        WarCry.stopChatListener()
-    end
-
-    -- write value to saved vars
-    WarCry.savedVariables.groupCommands = value
-end
-
-function WarCry.startChatListener()
-    EVENT_MANAGER:RegisterForEvent(WarCry.name, EVENT_CHAT_MESSAGE_CHANNEL, WarCry.chatCallback)
-    d("|c8B0000WarCry: group listener enabled|r")
-end
-
-function WarCry.stopChatListener()
-    EVENT_MANAGER:UnregisterForEvent(WarCry.name, EVENT_CHAT_MESSAGE_CHANNEL)
-    d("|c8B0000WarCry: group listener disabled|r")
-end
+end ]]
 
 
 --[[
@@ -79,25 +33,3 @@ function WarCry:Init()
     EVENT_MANAGER:UnregisterForEvent("WarCry.name", EVENT_PLAYER_ACTIVATED)
 end
 ]]--
-
-function WarCry.OnAddOnLoaded(event, addonName)
-    if addonName == WarCry.name then
-        --EVENT_MANAGER:RegisterForEvent("WarCry.name", EVENT_PLAYER_ACTIVATED, self.Init)
-
-        WarCry.savedVariables = WarCry.savedVariables or {}
-        WarCry.savedVariables = ZO_SavedVars:NewAccountWide("WarCryVars", WarCry.variableVersion, nil, WarCry.defaultVariables, GetWorldName())
-
-        WarCry.icon = GetCollectibleIcon(WarCry.savedVariables.collectibleID)
-
-        if WarCry.savedVariables.groupCommands == true then
-            WarCry.startChatListener()
-        end
-    end
-end
-
-EVENT_MANAGER:RegisterForEvent(WarCry.name, EVENT_ADD_ON_LOADED, WarCry.OnAddOnLoaded)
-
-SLASH_COMMANDS["/ahu"] = function(str)
-    WarCry.Play()
-    return
-end