+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.0
+ - reworking the whole base - now the listening string is the identifier for the WarCry
+ - adding "roar"
+ - convert addon to a lib - will be finished in the next update
+
 # 0.2.0
  - adding group sync via chat. only the leader of the group is allowed to trigger the animation for the group members
  - adding ability to disable the chat listener